Exo versionning

+++ JobOverview/Entities/Travail.cs
@@ -5,6 +5,6 @@
         public DateOnly DateTravail { get; set; }
         public int IdTache { get; set; }
         public decimal Heures { get; set; }
-        public float TauxProductivite { get; set; }
+        public decimal TauxProductivite { get; set; }
     }
 }
